Add template for `~/.pg_service.conf`
The template will generate a PostgreSQL connection service file
using 'chezmoi'.  The list of instances is fetched from our 'vault'
and all necessary certificates are downloaded and referenced in the
service definitions. However, at the moment our vault doesn't contain
the individual databases and their credentials. Hence the services
don't work as such, but the template and script illustrate the idea…

To try use the following:

  $ brew install vault chezmoi
  $ vault login -method=oidc role=zon-sudo
  $ chezmoi init https://github.com/ZeitOnline/dotfiles.git
  $ chezmoi apply