feat(pg-services): write service definitions for all projects (#57)
* chore(pg-services): re-enable "pg services" file and remove Postico favorites

Refs: 7414c12c, 32e3fca4

* style(pg-services): simplify things and always use `secretJSON`

* chore(pg-services): use helper script to handle (permission) errors

The wrapper enables the template to not break when access to certain
secrets is missing. Please note that running `chezzon init` is required
to update the configured "secret command".

* feat(pg-services): write service definitions for all projects

* chore(pg-services): add comments stating projects and instances

* style(pg-services): save some white space

* style(pg-services): simplify a bit more

* style(pg-services): use separate directories for instance certificates

* fix(pg-services): fix error & output handling for `vault`

`vault` returns an error code both when the use has no permissions to
"list" the path as well as when it's empty. In the latter case it also
returns `{}` (which only makes so much sense). The script now handles
both cases.

* chore(pg-services): add comment that might be helpful later