feat: set up "automatic" 'pre-commit' hooks using git `templatedir`
This adds the regular ("pre-commit", for code linting) and "commit-msg"
(for commit message linting) hooks to a 'git' template directory and
configures 'git' to use it. This way whenever a repository is cloned
or created, it will already have these hooks set.

See https://pre-commit.com/#pre-commit-init-templatedir

Note that this is implemented using a "run" script instead of a
template, because the generated hooks contain references to the
currently used Python version, which automatically get updated
this way.

Refs: ZO-3152, 1c1845ba